# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_MeshCore-Beacon_beacon-server_internal_api.Channel:
    properties:
      channelHash:
        description: hex-encoded single-byte hash
        type: string
      hashtag:
        description: 'tag name without # prefix; non-nil only for hashtag channels'
        type: string
      id:
        type: integer
      isHashtag:
        description: true if key was derived from a hashtag PSK
        type: boolean
      keyFingerprint:
        description: first 8 bytes of SHA256(key), hex-encoded
        type: string
      keyKnown:
        description: true if Beacon has a decryption key for this channel
        type: boolean
      lastSeen:
        description: epoch ms, time of most recent message
        type: integer
      messageCount:
        type: integer
      name:
        description: display name from config or nil
        type: string
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ChannelMessage:
    properties:
      channelHash:
        description: hex-encoded single-byte channel hash
        type: string
      content:
        description: decrypted message text
        type: string
      id:
        type: integer
      observationCount:
        description: number of packet_observations rows for this message's packet
          hash
        type: integer
      packetHash:
        description: hex-encoded packet hash for correlation with packet events
        type: string
      senderName:
        description: display name from the decrypted payload
        type: string
      sentAt:
        description: epoch ms, from the sender's embedded timestamp
        type: integer
    type: object
  github_com_MeshCore-Beacon_beacon-server_internal_api.ChannelSummary:
    properties:
      channelHash:
        description: hex-encoded single-byte hash
        type: string
      id:
        type: integer
      isHashtag:
        description: true if key was derived from a hashtag PSK
        type: boolean
      keyKnown:
        description: true if Beacon has a decryption key for this channel
        type: boolean
      lastSeen:
        description: epoch ms, time of most recent message
        type: integer
      name:
        description: display name from config or nil
        type: string
    type: object
...
  github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_ChannelSummary:
    properties:
      hasMore:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ChannelSummary'
        type: array
      nextCursor:
        type: integer
    type: object
...
  internal_api_handlers.APIError:
    properties:
      code:
        description: e.g. "not_found", "bad_request"
        type: string
      message:
        description: e.g. "channel not found"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: MeshCore Beacon
    url: https://github.com/MeshCore-Beacon/beacon-server
  description: MeshCore network observation backend. Ingests LoRa packets from MQTT
    brokers, stores in PostgreSQL, and streams live events via WebSocket.
  license:
    name: AGPL-3-or-later
  termsOfService: https://github.com/MeshCore-Beacon/beacon-server
  title: MeshCore Beacon API
  version: 1.6.0
paths:
...
  /channels:
    get:
      parameters:
      - description: Single-byte channel hash (hex)
        in: query
        name: hash
        type: string
      - description: Filter by IATA code
        in: query
        name: iata
        type: string
      - description: Filter by IATA code(s), comma-separated e.g. YOW or YOW,YYZ
        in: query
        name: iatas
        type: string
      - description: last_seen epoch ms of last item for pagination
        in: query
        name: cursor
        type: integer
      - description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Page-github_com_MeshCore-Beacon_beacon-server_internal_api_ChannelSummary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List channels
      tags:
      - Channels
  /channels/{channelID}:
    get:
      parameters:
      - description: Channel integer ID
        in: path
        name: channelID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.Channel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Get channel detail
      tags:
      - Channels
  /channels/{channelID}/messages:
    get:
      parameters:
      - description: Channel integer ID
        in: path
        name: channelID
        required: true
        type: integer
      - description: Return messages after this epoch ms
        in: query
        name: since
        type: integer
      - description: Filter by IATA code(s), comma-separated e.g. YVR or YVR,YYJ
        in: query
        name: iatas
        type: string
      - description: Filter by region ID, expands to member IATAs
        in: query
        name: regionId
        type: integer
      - description: Filter by region slug, expands to member IATAs
        in: query
        name: region
        type: string
      - description: 'Filter by transport scope name e.g. %23bc (URL-encoded #bc)'
        in: query
        name: scope
        type: string
      - description: Message ID of last item for pagination (results ordered newest
          first)
        in: query
        name: cursor
        type: integer
      - description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List messages for a channel
      tags:
      - Channels
...
  /messages:
    get:
      parameters:
      - description: Filter by channel integer ID (mutually exclusive with channelHash)
        in: query
        name: channelID
        type: integer
      - description: Filter by channel hash byte hex (mutually exclusive with channelID)
        in: query
        name: channelHash
        type: string
      - description: Return messages after this epoch ms
        in: query
        name: since
        type: integer
      - description: Filter by IATA code(s), comma-separated e.g. YVR or YVR,YYJ
        in: query
        name: iatas
        type: string
      - description: Filter by region ID, expands to member IATAs
        in: query
        name: regionId
        type: integer
      - description: Filter by region slug, expands to member IATAs
        in: query
        name: region
        type: string
      - description: 'Filter by transport scope name e.g. %23bc (URL-encoded #bc)'
        in: query
        name: scope
        type: string
      - description: Message ID of last item for pagination (results ordered newest
          first)
        in: query
        name: cursor
        type: integer
      - description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: List channel messages
      tags:
      - Messages
  /messages/backfill:
    get:
      parameters:
      - description: Return messages after this ID (use last WS event message ID)
        in: query
        name: afterId
        required: true
        type: integer
      - description: Filter by IATA code(s), comma-separated
        in: query
        name: iatas
        type: string
      - description: Filter by region slug
        in: query
        name: region
        type: string
      - description: Filter by region ID
        in: query
        name: regionId
        type: integer
      - description: Filter by transport scope name
        in: query
        name: scope
        type: string
      - description: Max results (default 20, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_MeshCore-Beacon_beacon-server_internal_api.ChannelMessage'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api_handlers.APIError'
      summary: Backfill messages after a given message ID
      tags:
      - Messages
...
schemes:
- http
- https
swagger: "2.0"
...